backend/requirements.txt
aiohttp

httpx[http2]

openai

python-dotenv
gTTS
